+++ app/src/main/java/org/jellyfin/androidtv/playback/overlay/actions/SelectAudioAction.java
@@ -13,4 +13,6 @@ public class SelectAudioAction extends CustomAction {
         setIcon(drawable);
         setDrawables(new Drawable[]{drawable});
     }
+
+
 }
