+++ app/src/main/res/layout/vlc_player_interface.xml
@@ -40,7 +40,7 @@
 
         <fragment
             android:id="@+id/leanback_fragment"
-            android:name="org.jellyfin.androidtv.playback.LeanbackOverlayFragment"
+            android:name="org.jellyfin.androidtv.playback.overlay.LeanbackOverlayFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_gravity="center" />
